requirements_frozen.txt
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
cssselect2==0.8.0
Django==5.2.7
django-ckeditor-5==0.2.18
django-cors-headers==4.9.0
django-js-asset==3.1.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ffmpeg-python==0.2.0
freetype-py==2.5.1
future==1.0.0
gunicorn==23.0.0
html5lib==1.1
idna==3.11
lxml==6.0.2
mysqlclient==2.2.7
oscrypto==1.3.0
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
pycairo==1.28.0
pycparser==2.23
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
PyMySQL==1.1.2
pypdf==6.1.3
python-bidi==0.6.7
PyYAML==6.0.3
redis==7.0.0
reportlab==4.4.4
requests==2.32.5
rlPyCairo==0.4.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
typing_extensions==4.15.0
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
wheel==0.45.1
xhtml2pdf==0.2.17
